llama/requirements.txt
# LlamaIndex 核心库
llama-index>=0.9.0

# LlamaIndex 扩展
llama-index-llms-siliconflow
llama-index-embeddings-siliconflow
llama-index-graph-stores-neo4j

# 数据库
neo4j>=5.0.0

# 网络请求
aiohttp>=3.8.0
requests>=2.28.0

# 环境变量管理
python-dotenv>=1.0.0

# 文档处理
pypdf>=3.0.0  # PDF 文件支持
pymupdf>=1.23.0 # PDF 渲染支持 (更强大的图像提取)
python-docx>=0.8.11  # Word 文档支持
Pillow>=9.0.0

# 腾讯云 COS
cos-python-sdk-v5>=1.9.0

# Web 框架
Flask>=2.0.0
gunicorn>=20.1.0

# 工具库
Werkzeug>=2.0.0
